+++ gulpfile.js
@@ -9,10 +9,23 @@ gulp.task('build', function () {
     .pipe(babel())
     .pipe(sourcemaps.write('.'))
     .pipe(gulp.dest('dist'))
+});
+
+gulp.task('start', function () {
+  return gulp.src('src/**/*.js')
+    .pipe(sourcemaps.init())
+    .pipe(babel())
+    .pipe(sourcemaps.write('.'))
+    .pipe(gulp.dest('dist'))
     .pipe(livereload());
 });
 
 gulp.task('watch', function () {
+  livereload.listen();
+  gulp.watch('src/*.js', ['start']);
+});
+
+gulp.task('watch:b', function () {
   livereload.listen();
   gulp.watch('src/*.js', ['build']);
 });