+++ src/async/index.js
@@ -1,3 +1,4 @@
 import './map'
 import './val'
 import './value'
+import './mapReduce'